% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddIndicators.R
\name{AddIndicators}
\alias{AddIndicators}
\title{Returns a data.frame with appropriate pre/post indicators for a change in some indepdent variable.}
\usage{
AddIndicators(df, num.periods.post, num.periods.pre, independent.variable,
  grouping.variable, epsilon = 0.01)
}
\arguments{
\item{df}{data frame}

\item{num.periods.post}{how many post-change indicators are needed}

\item{num.periods.pre}{how many pre-change indicators are needed}

\item{independent.variable}{the indepenent variable}

\item{grouping.variable}{the experimental units (e.g., states, cities, people)}

\item{epsilon}{minimally detectable change in var.name}
}
\value{
A dataframe with indicators added.
}
\description{
Returns a data.frame with appropriate pre/post indicators for a change in some indepdent variable.
}
