% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CherryPickedOK.R
\name{CherryPickedOK}
\alias{CherryPickedOK}
\title{Detects whether there is a change in some vector "k" places away}
\usage{
CherryPickedOK(x, k, epsilon)
}
\arguments{
\item{x}{numerical vector of interest}

\item{k}{number of positions}

\item{epsilon}{minimally detectable change}
}
\value{
cv a vector of booleans for whether or not there was a change.
}
\description{
Detects whether there is a change in some vector "k" places away
}
