% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractCoef.R
\name{ExtractCoef}
\alias{ExtractCoef}
\title{Extracts coefficients from a model}
\usage{
ExtractCoef(m, num.periods.pre, num.periods.post, felm.model = FALSE)
}
\arguments{
\item{m}{model}

\item{num.periods.pre}{how many pre-period indicators are used}

\item{num.periods.post}{how many post-period indicators are used}

\item{felm.model}{whether it is a lfe/felm model}
}
\value{
data frame
}
\description{
Extracts coefficients from a model
}
