% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenFormula.R
\name{GenFormula}
\alias{GenFormula}
\title{Generates a formula for use in dynamic model}
\usage{
GenFormula(x, num.periods.pre, num.periods.post)
}
\arguments{
\item{x}{regressor of interst}

\item{num.periods.pre}{number of pre-periods}

\item{num.periods.post}{number of post periods}
}
\value{
a string that can be used as a formula
}
\description{
Generates a formula for use in dynamic model
}
