% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HasChange.R
\name{HasChange}
\alias{HasChange}
\title{For a given vector x, for each position, return whether the value at that position is different
from the position k places away, with "different" determined by epsilon. If there is no position
k places away, then there is no change.}
\usage{
HasChange(x, k, epsilon)
}
\arguments{
\item{x}{numerical vector of interest}

\item{k}{number of positions}

\item{epsilon}{minimally detectable change}
}
\value{
cv a vector of booleans for whether or not there was a change.
}
\description{
For a given vector x, for each position, return whether the value at that position is different
from the position k places away, with "different" determined by epsilon. If there is no position
k places away, then there is no change.
}
