% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumChanges.R
\name{SumChanges}
\alias{SumChanges}
\title{Number of changes k places away}
\usage{
SumChanges(x, k, epsilon)
}
\arguments{
\item{x}{numerical vector of interest}

\item{k}{number of positions}

\item{epsilon}{minimally detectable change}
}
\value{
v a vector of past changes
}
\description{
Number of changes k places away
}
