% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractCoef.R
\name{ExtractCoef}
\alias{ExtractCoef}
\title{Extracts coefficients from a model}
\usage{
ExtractCoef(m, num.periods.pre, num.periods.post, felm.model = FALSE)
}
\arguments{
\item{felm.model}{}

\item{model}{}

\item{number}{of periods pre}

\item{number}{of periods post}
}
\value{
data frame
}
\description{
Extracts coefficients from a model
}
